imtcp: Implement "requirebind" option
This option, when enabled, will cause rsyslogd to abort if none of the
configured imtcp listeners can be bound. This is helpful for use cases
where the only inputs to rsyslogd arrive via TCP; if such a server cannot
listen, then it cannot do anything useful.

Granted, this method of handling this scenario is a bit messy. Ideally we
would have a way to specify any input source (in any module) as "required"
and a way for the core to handle cases where such "required" sources
are unavailable.

Signed-off-by: John Brooks <jbrooks@ciena.com>